requirements.txt
requests
beautifulsoup4
plyer
pyyaml
